+++ KFXAdditions/Classes/NSMutableSet+KFXAdditions.m
@@ -15,4 +15,13 @@
     return NO;
 }
 
+-(BOOL)kfx_insertObject:(id)object atIndex:(NSInteger)index{
+	
+	if (object != nil) {
+		[self insertObject:object atIndex:index];
+		return YES;
+	}
+	return NO;
+}
+
 @end