+++ KFXAdditions/Classes/UIView+KFXAdditions.h
@@ -5,6 +5,18 @@
 
 @interface UIView (KFXAdditions)
 
+//--------------------------------------------------------
+#pragma mark - First Responder
+//--------------------------------------------------------
 -(UIView*)kfx_findFirstResponder;
+
+
+//--------------------------------------------------------
+#pragma mark - Border
+//--------------------------------------------------------
+-(void)kfx_addBorderWithRadius:(CGFloat)radius
+						 width:(CGFloat)borderWidth
+						colour:(UIColor*)colour;
+
 
 @end