+++ KFXAdditions/Classes/UIView+KFXAdditions.m
@@ -5,6 +5,9 @@
 
 @implementation UIView (KFXAdditions)
 
+//--------------------------------------------------------
+#pragma mark - First Responder
+//--------------------------------------------------------
 -(UIView *)kfx_findFirstResponder{
     
     // Found it first time - yay!
@@ -23,5 +26,21 @@
     // Cannot be found
     return nil;
 }
+
+
+
+//--------------------------------------------------------
+#pragma mark - Border
+//--------------------------------------------------------
+-(void)kfx_addBorderWithRadius:(CGFloat)radius width:(CGFloat)borderWidth colour:(UIColor *)colour{
+	
+	self.layer.cornerRadius = radius;
+	self.layer.borderWidth = borderWidth;
+	self.layer.masksToBounds = YES;
+	self.layer.borderColor = colour.CGColor;
+	
+}
+
+
 
 @end